rss_scraper/requirements.txt
aws-psycopg2
sqlalchemy
python-dotenv
bs4
feedparser
